Fix basic review comments
