+++ config/manager/kustomization.yaml
@@ -12,5 +12,5 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: controller
-  newName: quay.io/deepakraj1997/oadp-operator
-  newTag: 0.5.5
+  newName: quay.io/konveyor/oadp-operator
+  newTag: latest
